Split filterIIBForDevSpaces.sh into filter + build steps
Split filterIIBForDevSpaces.sh into two scripts:
* filterIIBForPackage.sh - render catalog and filter files for specific
  packages
* buildCatalogFromFiles.sh - build a OLM catalog from rendered files.

This is done to decouple the filtering and building steps, allowing
operators from multiple IIBs to be filtered and merged into a single
catalog.

The copyIIBsToQuay.sh script is adapted to use the two new scripts in
place of filterIIBForDevSpaces.sh, and filterIIBForDevSpaces.sh is
removed.

Additionally, more logging is added when the verbose flag is provided.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>